Moved lang synopsis, and some revisions
