+++ src/StoreKeeper/WooCommerce/B2C/Commands/AbstractCommand.php
@@ -11,6 +11,7 @@ use StoreKeeper\WooCommerce\B2C\Core;
 use StoreKeeper\WooCommerce\B2C\Exceptions\BaseException;
 use StoreKeeper\WooCommerce\B2C\Exceptions\NotConnectedException;
 use StoreKeeper\WooCommerce\B2C\Exceptions\SubProcessException;
+use StoreKeeper\WooCommerce\B2C\I18N;
 use StoreKeeper\WooCommerce\B2C\Options\StoreKeeperOptions;
 use StoreKeeper\WooCommerce\B2C\Tools\StoreKeeperApi;
 
@@ -63,7 +64,7 @@ abstract class AbstractCommand implements CommandInterface
 
     protected static function generateExamples(array $examples): string
     {
-        $exampleBuilder = "\n\n## EXAMPLES\n\n";
+        $exampleBuilder = "\n\n## ".__('EXAMPLES', I18N::DOMAIN)."\n\n";
         $exampleBuilder .= implode("\n\n", $examples);
 
         return $exampleBuilder;
