+++ src/StoreKeeper/WooCommerce/B2C/Commands/FileExports/AbstractFileExportCommand.php
@@ -3,6 +3,7 @@
 namespace StoreKeeper\WooCommerce\B2C\Commands\FileExports;
 
 use StoreKeeper\WooCommerce\B2C\Commands\AbstractCommand;
+use StoreKeeper\WooCommerce\B2C\I18N;
 use StoreKeeper\WooCommerce\B2C\Interfaces\IFileExportCommand;
 use StoreKeeper\WooCommerce\B2C\Tools\Language;
 
@@ -13,6 +14,20 @@ abstract class AbstractFileExportCommand extends AbstractCommand implements IFil
         'nl',
         'de',
     ];
+
+    public static function getSynopsis(): array
+    {
+        return [
+            [
+                'type' => 'assoc',
+                'name' => 'lang',
+                'description' => __('The language to which the entities will be exported.', I18N::DOMAIN),
+                'optional' => true,
+                'default' => Language::getSiteLanguageIso2(),
+                'options' => self::LANGUAGE_OPTIONS,
+            ],
+        ];
+    }
 
     public function execute(array $arguments, array $assoc_arguments)
     {
