+++ src/StoreKeeper/WooCommerce/B2C/Commands/ProcessAllTasks.php
@@ -45,6 +45,12 @@ class ProcessAllTasks extends AbstractCommand
                 'optional' => true,
                 'default' => 0,
             ],
+            [
+                'type' => 'flag',
+                'name' => WpCliCommandRunner::SINGLE_PROCESS,
+                'description' => __('Flag to prevent spawning of child processes. Having this might cause timeouts during execution.', I18N::DOMAIN),
+                'optional' => true,
+            ],
         ];
     }
