+++ src/StoreKeeper/WooCommerce/B2C/Commands/SyncWoocommerceCrossSellProducts.php
@@ -29,6 +29,12 @@ class SyncWoocommerceCrossSellProducts extends AbstractSyncCommand
                 'description' => __('Specify total amount of cross-sell products from Storekeeper Backoffice. By default, counts the total amount of products by checking the Storekeeper Backoffice', I18N::DOMAIN),
                 'optional' => true,
             ],
+            [
+                'type' => 'flag',
+                'name' => WpCliCommandRunner::SINGLE_PROCESS,
+                'description' => __('Flag to prevent spawning of child processes. Having this might cause timeouts during execution.', I18N::DOMAIN),
+                'optional' => true,
+            ],
         ];
     }
