+++ src/StoreKeeper/WooCommerce/B2C/Commands/SyncWoocommerceProducts.php
@@ -25,6 +25,12 @@ class SyncWoocommerceProducts extends AbstractSyncCommand
                 'description' => __('Specify total amount of products from Storekeeper Backoffice. By default, counts the total amount of products by checking the Storekeeper Backoffice', I18N::DOMAIN),
                 'optional' => true,
             ],
+            [
+                'type' => 'flag',
+                'name' => WpCliCommandRunner::SINGLE_PROCESS,
+                'description' => __('Flag to prevent spawning of child processes. Having this might cause timeouts during execution.', I18N::DOMAIN),
+                'optional' => true,
+            ],
         ];
     }
